+++ WebApp/frontend/public/favicon.ico
