+++ WebApp/frontend/public/manifest.json
