+++ WebApp/frontend/src/Assets/GlobleStyles/GlobleStyles.scss
@@ -6,7 +6,6 @@
     --text-color: #333;
 
     --primary: rgb(98, 199, 227);
-
     --header-height: 70px ;
     --sidebar-width: 288px ;
