+++ WebApp/frontend/src/Layouts/MainLayout/MainLayout.scss
