+++ WebApp/frontend/src/routes/index.js
@@ -1,0 +1,8 @@
+import Dashboard from "../Pages/Dashborad"
+import MainLayout from "../Layouts/MainLayout"
+
+export const routes = [
+    {
+        path: '/' , component: Dashboard, layout: MainLayout
+    },
+]
