--- web-app/src/routes/index.js
@@ -1,8 +1,0 @@
-import Dashboard from "../Pages/Dashboard"
-import MainLayout from "../Layouts/MainLayout"
-
-export const routes = [
-    {
-        path: '/dashboard', component: Dashboard, layout: MainLayout
-    },
-]
